Undo change file structure
