+++ src/reanimated2/animation/decay/decay.ts
@@ -1,8 +1,5 @@
 'use strict';
-import {
-  defineAnimation,
-  getReduceMotionForAnimation,
-} from '../defineAnimation';
+import { defineAnimation, getReduceMotionForAnimation } from '../utils';
 import type {
   AnimationCallback,
   Timestamp,
